pyproject.toml
[project]
name = "nbquiz"
authors = [
    {name = "Mike Matera", email = "michael.matera@cabrillo.edu"},
]
version = "0.3.0"
requires-python = ">= 3.10"
dependencies = [
    "nbtest@git+https://github.com/mike-matera/nbtest.git#egg=main",
    "jinja2", 
    "jupyterlab",
    "ipywidgets",
    "grpcio",
    "grpcio-tools",
    "myst_parser",
    "docutils",
    "cssutils",
    "minify_html",
    "pyyaml",
    "yamale",
    "requests",
]

[tool.setuptools.package-data]
"nbquiz.resources" = ["*"]
"nbquiz.resources.canvas" = ["*"]
"nbquiz.resources.canvas.assessment" = ["*"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.scripts]
nbquiz = "nbquiz.cli:main"

[tool.ruff]
line-length = 79

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
